[SQUASH] Merge tag 'LA.UM.9.12.r1-11500-SMxx50.0' of https://source.codeaurora.org/quic/la/platform/hardware/qcom/display into lineage-19.0-caf-sm8250
"LA.UM.9.12.r1-11500-SMxx50.0"

* tag 'LA.UM.9.12.r1-11500-SMxx50.0' of https://source.codeaurora.org/quic/la/platform/hardware/qcom/display:
  display: Allow Idle fallback
  sdm: Add support for allowIdleFallback interface
  composer: Add DisplayConfig query for RC support
  hwc2 : Add scaling rgb layer flag to layer stack
  sdm: Fix rounded corner functionality
  sdm : Add support for rounded corner
  composer: fix the dependency on IDC 2.0
  hwc2 : Add scaling rgb layer flag to layer stack
  composer: Fix corruption issue due to wrong index
  composer: Add support for camera smooth notification
  display: fix compilation for 5.4 kernel
  composer: Return correct CONFIG_GROUP when queried
  libdrmutils: Fix compilation for 5.4 kernel
  composer: Add support for TWM entry and exit.
  composer: Pass hwc tid to perf-hal module
  composer:set perf-hint for large composition cycle
  composer: Remove kVsyncTimeDriftNs
  composer: Fix async power mode issue
  sdm: Add dtrace_end to SetVsyncStatus.
  sdm: Do not set min fps during idling if qsync is enabled.
  sdm: Add support for qsync based idle fallback.
  composer: Get vsync period in dummy display
  sdm: add drm format for BGR888
  gralloc: fix add-overflow in ALIGN API
  sdm: add GetConfig method for built-in displays
  composer: avoid skip flag for game layers
  composer: reset secure session active bit when session ends
  Revert "composer: reset validate on camera launch"
  sdm: Fix async power mode issues

 Conflicts:
        sdm/include/core/layer_stack.h

+++ composer/QtiComposerClient.cpp
@@ -1878,7 +1878,7 @@ bool QtiComposerClient::CommandReader::parseSetLayerPerFrameMetadataBlobs(uint16
   for (const auto& m : metadata) {
     keys.push_back(static_cast<int32_t>(m.key));
     sizes_of_metablob_.push_back(m.blob.size());
-    for (uint8_t i = 0; i < m.blob.size(); i++) {
+    for (size_t i = 0; i < m.blob.size(); i++) {
       blob_of_data_.push_back(m.blob[i]);
     }
   }
